Menu Update
Created a better menu for the program.

Got rid of the idea to view board game data from the main menu because it was a bit redundant. Viewing the text files in notepad is simply better than doing it within the program.

Includes a file with tracked information on the game summoner wars as an example file.